v2/requirements.txt
# V2 Trading Bot Requirements
# ================================

# Environment
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Async HTTP
aiohttp>=3.9.0

# WebSocket
websockets>=12.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Parquet Storage
pyarrow>=14.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging (colorful)
colorama>=0.4.6
